+++ server/protocol.js
@@ -84,7 +84,7 @@ export const MessageType = Object.freeze({
 });
 
 // Versione
-export const VERSION               = '1.14.0';
+export const VERSION               = '1.15.0';
 
 // Configurazione
 export const DEFAULT_PORT          = 8765;